Ahmadhi

See merge request AdvProg/reguler-2023/mahasiswa/kelas-a/2106629963-Muflih-Naufal-Maxi-/kang-warnet/bayar-service!26

+++ src/main/java/id/ac/ui/cs/advprog/bayarservice/dto/Invoice/InvoiceRequest.java
@@ -25,10 +25,10 @@ public class InvoiceRequest {
 
     @Min(0)
     @NotNull(message = "totalAmount is mandatory")
-    private Integer totalAmount;
+    private Long totalAmount;
 
     @Min(0)
-    private Integer discount;
+    private Long discount;
 
     public Invoice toEntity() {
         Invoice entity = new Invoice();
